+++ runtime/src/main/java/hosh/runtime/prompt/HostnamePromptProvider.java
@@ -21,34 +21,21 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package hosh.runtime;
+package hosh.runtime.prompt;
 
-import hosh.runtime.prompt.PromptProvider;
 import hosh.spi.State;
-import org.jline.reader.EndOfFileException;
-import org.jline.reader.LineReader;
-import org.jline.reader.UserInterruptException;
 
-import java.util.Optional;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 
-public class ReplReader {
+public class HostnamePromptProvider implements PromptProvider {
 
-	private final PromptProvider promptProvider;
-	private final LineReader lineReader;
-
-	public ReplReader(PromptProvider promptProvider, LineReader lineReader) {
-		this.promptProvider = promptProvider;
-		this.lineReader = lineReader;
-	}
-
-	public Optional<String> read(State state) {
+	@Override
+	public String provide(State state) {
 		try {
-			String line = lineReader.readLine(promptProvider.provide(state));
-			return Optional.of(line);
-		} catch (UserInterruptException e) {
-			return Optional.of("");
-		} catch (EndOfFileException e) {
-			return Optional.empty();
+			return InetAddress.getLocalHost().getHostName();
+		} catch (UnknownHostException e) {
+			return null;
 		}
 	}
 }